+++ src/helpers.ts
@@ -1,5 +1,6 @@
-const _minimist = require('minimist');
+import _minimist = require('minimist');
 const _minimistString = require('minimist-string');
+const x2jsCore = require('x2js');
 
 export function isWindows(): boolean {
     return /win32|mswin(?!ce)|mingw|bccwin|cygwin/i.test(process.platform);
@@ -17,4 +18,11 @@ export function minimist(...x) {
 
 export function minimistString(...x) {
     return _minimistString(...x);
+}
+
+const x2js = new x2jsCore();
+export function xml2js(content): Promise<any> {
+    return new Promise(resolve => {
+        resolve(x2js.xml2js(content));
+    });
 }