+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the "vscode-phpunit" extension will be documented in this
 Check [Keep a Changelog](http://keepachangelog.com/) for 
 recommendations on how to structure this file.
 
+## 0.4.1
+- refactoring TextLine
+
 ## 0.4.0
 - support teamcity log
 - warning phpunit could not find